Added BoardDisplay Add new game components and functionality
This commit introduces a lot of changes primarily aimed at adding crucial components for the Snake game and improving its functionality.

A new GameObject with components such as SpriteRenderer and Transform was added, represented as a Unity YAML file. This prefab will serve as the tile for the game board.

The shape and appearance of the snake, food and empty cells are defined with three Material files (Snake.mat, Food.mat, NoneMaterial.mat).

Improvements were made on the GameManager script to initialise the game board with random positions for the snake and the food. Also, it was a refactoring of the Board class, moving it into the Snake.Scripts namespace and changing the enum members' names for better readability.

Another significant addition is the Scripts for game components like the BoardDisplay, which is an abstract class for displaying the board, and SpriteBoardDisplay, which is a specific implementation of BoardDisplay using SpriteRenderer for the cells.

TileDisplay script has been added to handle the appearance of individual tiles, and ColorChanger script has been to change the colors of the sprites using predefined materials.

In general, these changes provide fundamental components and increase the game functionality, allowing the further development of the game.