"Implemented multiple players support in Snake Game"
The implemented support for multiple players in the classic Snake Game. A Player struct was created containing snakeID and inputSchemer. This allows for the allocation of multiple snakes to different players. The handling of input was updated to accommodate different input schemes for each allocated snake. The changes also include minor fixes and organizational changes of the code to improve readability.