fix: snake crash on start
This commit fixes a bug where the game would crash on start due to the snake's direction attribute being set to Vector2Int.zero (a direction that doesn't exist).

To mitigate this, checks have been added to the inputHandler in the GameManager and InputController scripts to ensure inputDirection is not Vector2Int.zero before it's passed as NextDirection for the snake. The default nextDirection value in InputController has also been changed from currentDirection to default (Vector2Int.zero) for consistency.