Optimize Board reset in SnakeGame
to minimize the memory footprint and improve the runtime efficiency, changes were made in the reset method of the Board class.

Instead of instantiating a new List and array of Tiles everytime the board resets, the existing data structures are now cleared and reused. This prevents redundant memory allocation and garbage collection, leading to smoother performance. Also, by setting snake field to null explicitly, we ensure the memory of previous state is released promptly.