fix: Tiles still having a Tile.Snake reference after clearing the board
The code for the 'ClearBoard()' function in 'Board.cs' was updated to set 'Tiles[x, y].Snake' to null when clearing the board. This is necessary because previously although the tiles were being marked as empty, a reference to the snake still existed on the tile which could result in unexpected behaviour in gameplay. The same has been addressed now.

TL;DR: Tiles were retaining a reference to the snake even after being cleared, so an update was made to properly nullify this.